Module brontes_pricing::exchanges::uniswap_v3::batch_request

Modules

IGetUniswapV3PoolDataBatchRequest

Generated by the following Solidity interface…

IGetUniswapV3TickDataBatchRequest

Generated by the following Solidity interface…

ISyncUniswapV3PoolBatchRequest

Generated by the following Solidity interface…

Structs

PoolData

TickData

data_constructorCall

Function with signature data_constructor(address[]) and selector 0x45ef8379.

data_constructorReturn

Container type for the return parameters of the data_constructor(address[]) function.

tick_constructorCall

Function with signature tick_constructor(address,bool,int24,uint16,int24) and selector 0xc5314b8c.

tick_constructorReturn

Container type for the return parameters of the tick_constructor(address,bool,int24,uint16,int24) function.

Functions

get_uniswap_v3_tick_data_batch_request

get_v3_pool_data_batch_request