Module brontes_pricing::protocols::uniswap_v2::IUniswapV2Pair

Module containing a contract’s types and functions.

interface IUniswapV2Pair {
    function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
    function token0() external view returns (address);
    function token1() external view returns (address);
    function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes calldata data);
    event Sync(uint112 reserve0, uint112 reserve1);
}

Structs

Sync

Event with signature Sync(uint112,uint112) and selector 0x1c411e9a96e071241c2f21f7726b17ae89e3cab4c78be50e062b03a9fffbbad1.

getReservesCall

Function with signature getReserves() and selector 0x0902f1ac.

getReservesReturn

Container type for the return parameters of the getReserves() function.

swapCall

Function with signature swap(uint256,uint256,address,bytes) and selector 0x022c0d9f.

swapReturn

Container type for the return parameters of the swap(uint256,uint256,address,bytes) function.

token0Call

Function with signature token0() and selector 0x0dfe1681.

token0Return

Container type for the return parameters of the token0() function.

token1Call

Function with signature token1() and selector 0xd21220a7.

token1Return

Container type for the return parameters of the token1() function.

Enums

IUniswapV2PairCalls

Container for all the IUniswapV2Pair function calls.

IUniswapV2PairEvents

Container for all the IUniswapV2Pair events.