Module brontes_pricing::protocols::uniswap_v3::IUniswapV3Pool

Module containing a contract’s types and functions.

interface IUniswapV3Pool {
    function token0() external view returns (address);
    function token1() external view returns (address);
    function liquidity() external view returns (uint128);
    function slot0() external view returns (uint160, int24, uint16, uint16, uint16, uint8, bool);
    function fee() external view returns (uint24);
    function tickSpacing() external view returns (int24);
    function ticks(int24 tick) external view returns (uint128, int128, uint256, uint256, int56, uint160, uint32, bool);
    function tickBitmap(int16 wordPosition) external view returns (uint256);
    function swap(address recipient, bool zeroForOne, int256 amountSpecified, uint160 sqrtPriceLimitX96, bytes calldata data) external returns (int256, int256);
    event Swap(address indexed sender, address indexed recipient, int256 amount0, int256 amount1, uint160 sqrtPriceX96, uint128 liquidity, int24 tick);
    event Burn(address indexed owner, int24 indexed tickLower, int24 indexed tickUpper, uint128 amount, uint256 amount0, uint256 amount1);
    event Mint(address sender, address indexed owner, int24 indexed tickLower, int24 indexed tickUpper, uint128 amount, uint256 amount0, uint256 amount1);
}

Structs

Burn

Event with signature Burn(address,int24,int24,uint128,uint256,uint256) and selector 0x0c396cd989a39f4459b5fa1aed6a9a8dcdbc45908acfd67e028cd568da98982c.

Mint

Event with signature Mint(address,address,int24,int24,uint128,uint256,uint256) and selector 0x7a53080ba414158be7ec69b987b5fb7d07dee101fe85488f0853ae16239d0bde.

Swap

Event with signature Swap(address,address,int256,int256,uint160,uint128,int24) and selector 0xc42079f94a6350d7e6235f29174924f928cc2ac818eb64fed8004e115fbcca67.

feeCall

Function with signature fee() and selector 0xddca3f43.

feeReturn

Container type for the return parameters of the fee() function.

liquidityCall

Function with signature liquidity() and selector 0x1a686502.

liquidityReturn

Container type for the return parameters of the liquidity() function.

slot0Call

Function with signature slot0() and selector 0x3850c7bd.

slot0Return

Container type for the return parameters of the slot0() function.

swapCall

Function with signature swap(address,bool,int256,uint160,bytes) and selector 0x128acb08.

swapReturn

Container type for the return parameters of the swap(address,bool,int256,uint160,bytes) function.

tickBitmapCall

Function with signature tickBitmap(int16) and selector 0x5339c296.

tickBitmapReturn

Container type for the return parameters of the tickBitmap(int16) function.

tickSpacingCall

Function with signature tickSpacing() and selector 0xd0c93a7c.

tickSpacingReturn

Container type for the return parameters of the tickSpacing() function.

ticksCall

Function with signature ticks(int24) and selector 0xf30dba93.

ticksReturn

Container type for the return parameters of the ticks(int24) function.

token0Call

Function with signature token0() and selector 0x0dfe1681.

token0Return

Container type for the return parameters of the token0() function.

token1Call

Function with signature token1() and selector 0xd21220a7.

token1Return

Container type for the return parameters of the token1() function.

Enums

IUniswapV3PoolCalls

Container for all the IUniswapV3Pool function calls.

IUniswapV3PoolEvents

Container for all the IUniswapV3Pool events.